+++ python/segmnist/__init__.py
@@ -1,7 +1,7 @@
 from segmnist.segmnistshapes import SegMNISTShapes
 # from fgmodattendexperiment import FGModAttendExperiment
-import loader
-from stimsetdispatcher import StimSetDispatcher
+from segmnist import loader
+from segmnist.stimsetdispatcher import StimSetDispatcher
 
 __all__ = ['SegMNISTShapes', 'FGModAttendExperiment', 'loader',
            'StimSetDispatcher']
