+++ python/segmnist/loader/mnist.py
@@ -5,33 +5,6 @@ import struct
 import numpy as np
 
 
-def load_standard_MNIST(name, shuffle, path=None):
-    if path is None:
-        if "MNIST_PATH" in os.environ:
-            path = os.environ["MNIST_PATH"]
-        else:
-            raise RuntimeError('Environment variable MNIST_PATH or ' +
-                                'function parameter path must be defined')
-
-    if shuffle:
-        from shuffled_mnist import ShuffledMNIST
-        D = ShuffledMNIST
-    else:
-        D = MNIST
-
-    if name == 'training' or name == 'mnist-training':
-        mnist = D(path, dataset_slice=(0, 5000))
-        mnist.load_standard('training')
-        return mnist
-    elif name == 'validation' or name == 'mnist-validation':
-        mnist = D(path, dataset_slice=(5000, 6000))
-        mnist.load_standard('training')
-        return mnist
-    else:
-        raise RuntimeError('Unknown standard MNIST-type dataset: %s' %
-                            name)
-
-
 class MNIST(object):
     """ Class to read MNIST dataset and iterate examples in original order.
     """
