+++ python/segmnist/pytorch/__init__.py
@@ -1,2 +1,1 @@
-from pytorch.segmnist_pytorch import SegMNISTShapesPyTorch
-# from .segmnist_pytorch import SegMNISTShapesPyTorch
+from segmnist.pytorch.segmnist_pytorch import SegMNISTShapesPyTorch
