+++ python/segmnist/textures/__init__.py
@@ -1,15 +1,9 @@
-from whitenoise import WhiteNoiseTexture
-# from textures.whitenoise import WhiteNoiseTexture
-from sinusoidal import SinusoidalGratings
-# from textures.sinusoidal import SinusoidalGratings
-from fgmod import FGModTexture
-# from textures.fgmod import FGModTexture
-from generator import TextureGenerator
-# from textures.generator import TextureGenerator
-from dispatcher import TextureDispatcher
-# from textures.dispatcher import TextureDispatcher
-from intermix import IntermixTexture
-# from textures.intermix import IntermixTexture
+from segmnist.textures.whitenoise import WhiteNoiseTexture
+from segmnist.textures.sinusoidal import SinusoidalGratings
+from segmnist.textures.fgmod import FGModTexture
+from segmnist.textures.generator import TextureGenerator
+from segmnist.textures.dispatcher import TextureDispatcher
+from segmnist.textures.intermix import IntermixTexture
 
 __all__ = ['WhiteNoiseTexture',
            'SinusoidalGratings',
